requirements.txt
torch
torchvision
streamlit
numpy
pillow
opencv-python
matplotlib
scikit-learn
seaborn
tqdm
